+++ trader/config/database.go
@@ -1,0 +1,47 @@
+package config
+
+import (
+	"database/sql"
+	"fmt"
+	"os"
+
+	_ "github.com/go-sql-driver/mysql"
+)
+
+var DB *sql.DB
+
+var (
+	MYSQL_USER            = os.Getenv("MYSQL_USER")
+	MYSQL_PASSWORD        = os.Getenv("MYSQL_PASSWORD")
+	MYSQL_HOST            = os.Getenv("MYSQL_HOST")
+	MYSQL_PORT            = os.Getenv("MYSQL_PORT")
+	MYSQL_DATABASE        = os.Getenv("MYSQL_DATABASE")
+	MYSQL_CONNECTION_NAME = os.Getenv("MYSQL_CONNECTION_NAME")
+)
+
+func DSN() string {
+	socketDir, isSet := os.LookupEnv("DB_SOCKET_DIR")
+	if !isSet {
+		socketDir = "/cloudsql"
+	}
+
+	var dsn string
+	if MYSQL_CONNECTION_NAME == "" {
+		dsn = fmt.Sprintf("%s:%s@tcp(%s:%s)/%s", MYSQL_USER, MYSQL_PASSWORD, MYSQL_HOST, MYSQL_PORT, MYSQL_DATABASE)
+	} else {
+		dsn = fmt.Sprintf("%s:%s@unix(/%s/%s)/%s", MYSQL_USER, MYSQL_PASSWORD, socketDir, MYSQL_CONNECTION_NAME, MYSQL_DATABASE)
+	}
+
+	return dsn
+}
+
+func init() {
+	dsn := DSN()
+	fmt.Println("dsn:", dsn)
+
+	var err error
+	DB, err = sql.Open("mysql", dsn)
+	if err != nil {
+		panic(err.Error())
+	}
+}
